[Terraform] Autogen resource manager operations
