+++ google-beta/resource_google_folder.go
@@ -63,7 +63,13 @@ func resourceGoogleFolderCreate(d *schema.ResourceData, meta interface{}) error 
 		return fmt.Errorf("Error creating folder '%s' in '%s': %s", displayName, parent, err)
 	}
 
-	err = resourceManagerV2Beta1OperationWait(config.clientResourceManager, op, "creating folder")
+	opV1 := make(map[string]interface{})
+	waitErr := Convert(op, opV1)
+	if waitErr != nil {
+		return waitErr
+	}
+
+	err = resourceManagerOperationWaitTime(config, opV1, "", "creating folder", int(d.Timeout(schema.TimeoutCreate).Minutes()))
 
 	if err != nil {
 		return fmt.Errorf("Error creating folder '%s' in '%s': %s", displayName, parent, err)
@@ -132,7 +138,13 @@ func resourceGoogleFolderUpdate(d *schema.ResourceData, meta interface{}) error 
 			return fmt.Errorf("Error moving folder '%s' to '%s': %s", displayName, newParent, err)
 		}
 
-		err = resourceManagerV2Beta1OperationWait(config.clientResourceManager, op, "move folder")
+		opV1 := make(map[string]interface{})
+		waitErr := Convert(op, opV1)
+		if waitErr != nil {
+			return waitErr
+		}
+
+		err = resourceManagerOperationWaitTime(config, opV1, "", "move folder", int(d.Timeout(schema.TimeoutCreate).Minutes()))
 		if err != nil {
 			return fmt.Errorf("Error moving folder '%s' to '%s': %s", displayName, newParent, err)
 		}
